backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx[http2]>=0.27.0
PyJWT[crypto]>=2.8.0
pydantic-settings>=2.2.1
python-multipart>=0.0.9
itsdangerous>=2.2.0
firebase-admin>=6.6.0
pytest>=8.0.0
reportlab>=4.0.0
Pillow>=10.0.0
qrcode[pil]>=7.4.2
